Create post_waterlevel.py
Python script to fire HC-SR04 ultrasonic sensor and get distance reading.  Post this reading to waterlevels api for trending and monitoring.  The distance represents the distance from the top of the sump pit, where the sensor is mounted, to the surface of the water.  The smaller the number, the closer the water is to the top of the pit, potentially representing a pump failure.